requirements.txt
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
DateTime==5.2
Flask==2.3.2
Flask-Cors==4.0.0
gunicorn==21.2.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.6
zipp==3.16.2
zope.interface==6.0
